# offchain/backend-go/docs/swagger.yaml
basePath: /v1
definitions:
...
  github_com_vndc_backend_internal_models.ChallengeData:
    properties:
      expires_at:
        example: "2026-04-24T07:05:00Z"
        type: string
      message:
        description: Message is the full EIP-191 formatted message the wallet must
          sign.
        example: |-
          vndc.io wants you to sign in with your Ethereum account:
          0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045

          Sign in to VNDC Platform

          Nonce: a1b2c3d4e5f6...
        type: string
      nonce:
        description: Nonce is the single-use random string embedded in the message.
        example: a1b2c3d4e5f6a7b8c9d0e1f2
        type: string
    type: object
  github_com_vndc_backend_internal_models.ChallengeResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.ChallengeData'
      meta:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.ResponseMeta'
      success:
        example: true
        type: boolean
    type: object
  github_com_vndc_backend_internal_models.Complete2FARequest:
    properties:
      code:
        description: Code is a 6-digit TOTP code or an 8-character backup code.
        example: "123456"
        type: string
      device_name:
        example: iPhone 15 Pro
        type: string
      device_os:
        example: iOS 17
        type: string
      temp_token:
        description: TempToken is the value returned by POST /auth/login when requires_2fa
          is true.
        example: 9f8e7d6c5b4a3210fedcba9876543210
        type: string
    type: object
  github_com_vndc_backend_internal_models.Disable2FARequest:
    properties:
      code:
        description: Code is a 6-digit TOTP code or an 8-character backup code.
        example: "123456"
        type: string
    type: object
  github_com_vndc_backend_internal_models.Enable2FARequest:
    properties:
      code:
        description: Code is the current 6-digit TOTP code to confirm the secret is
          correctly configured.
        example: "123456"
        type: string
    type: object
  github_com_vndc_backend_internal_models.ErrorBody:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      details: {}
      message:
        example: wallet query param required
        type: string
    type: object
  github_com_vndc_backend_internal_models.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorBody'
      meta:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.ResponseMeta'
      success:
        example: false
        type: boolean
    type: object
  github_com_vndc_backend_internal_models.KYCDocumentInfo:
    properties:
      document_ref:
        example: enc:s3::bucket/usr123/passport.enc
        type: string
      review_note:
        example: Document blurry — please resubmit
        type: string
      reviewed_at:
        type: string
      submitted_at:
        example: "2026-04-24T07:00:00Z"
        type: string
      type:
        example: PASSPORT
        type: string
    type: object
  github_com_vndc_backend_internal_models.Login2FAData:
    properties:
      message:
        example: TOTP code required — call /auth/2fa/complete
        type: string
      requires_2fa:
        example: true
        type: boolean
      temp_token:
        description: TempToken is a short-lived (5 min) opaque token passed to /auth/2fa/complete.
        example: 9f8e7d6c5b4a3210fedcba9876543210
        type: string
    type: object
  github_com_vndc_backend_internal_models.Login2FAResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.Login2FAData'
      meta:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.ResponseMeta'
      success:
        example: true
        type: boolean
    type: object
  github_com_vndc_backend_internal_models.LoginRequest:
    properties:
      device_name:
        example: iPhone 15 Pro
        type: string
      device_os:
        example: iOS 17
        type: string
      signature:
        description: Signature is the 0x-prefixed EIP-191 personal_sign output.
        example: 0x4a5b6c...130hexchars...ab
        type: string
      wallet:
        description: Wallet is the checksummed Ethereum wallet address.
        example: 0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045
        type: string
    type: object
  github_com_vndc_backend_internal_models.LoginResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.TokenPairData'
      meta:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.ResponseMeta'
      success:
        example: true
        type: boolean
    type: object
  github_com_vndc_backend_internal_models.MessageResponse:
    properties:
      message:
        example: Operation completed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  github_com_vndc_backend_internal_models.RefreshRequest:
    properties:
      refresh_token:
        description: RefreshToken is the opaque token issued at login.
        example: b3c4d5e6f7a8b9c0d1e2f3a4b5c6d7e8...
        type: string
    type: object
...
  github_com_vndc_backend_internal_models.ResponseMeta:
    properties:
      request_id:
        example: 01HWXYZ1234567890ABCDEF
        type: string
      timestamp:
        example: "2026-04-24T07:00:00Z"
        type: string
    type: object
  github_com_vndc_backend_internal_models.SessionInfo:
    properties:
      created_at:
        example: "2026-04-24T07:00:00Z"
        type: string
      device_name:
        example: iPhone 15 Pro
        type: string
      device_os:
        example: iOS 17
        type: string
      expires_at:
        example: "2026-05-01T07:00:00Z"
        type: string
      geo_country:
        example: VN
        type: string
      id:
        example: 9f8e7d6c-5b4a-3210-fedc-ba9876543210
        type: string
      ip_address:
        example: 203.0.113.42
        type: string
      issued_at:
        example: "2026-04-24T07:00:00Z"
        type: string
      last_used_at:
        type: string
      roles:
        example:
        - '["USER"]'
        items:
          type: string
        type: array
      user_agent:
        example: Mozilla/5.0 ...
        type: string
      user_id:
        example: 65f1a2b3c4d5e6f7a8b9c0d1
        type: string
      wallet_address:
        example: 0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045
        type: string
    type: object
  github_com_vndc_backend_internal_models.SessionListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_vndc_backend_internal_models.SessionInfo'
        type: array
      meta:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.ResponseMeta'
      success:
        example: true
        type: boolean
    type: object
  github_com_vndc_backend_internal_models.Setup2FAData:
    properties:
      backup_codes:
        description: BackupCodes are one-time plaintext codes — shown ONCE. Store
          securely.
        example:
        - '["ABCD1234"'
        - '"EFGH5678"]'
        items:
          type: string
        type: array
      otp_auth_uri:
        description: OTPAuthURI is the otpauth:// URI to encode as a QR code.
        example: otpauth://totp/vndc.io:0xd8dA...?secret=JBSWY3DP&issuer=vndc.io
        type: string
      secret:
        description: Secret is the base-32 encoded TOTP secret to enter in an authenticator
          app.
        example: JBSWY3DPEHPK3PXP
        type: string
    type: object
  github_com_vndc_backend_internal_models.Setup2FAResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.Setup2FAData'
      meta:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.ResponseMeta'
      success:
        example: true
        type: boolean
    type: object
...
  github_com_vndc_backend_internal_models.TokenPairData:
    properties:
      access_token:
        description: 'AccessToken is a short-lived JWT (15 min). Send in Authorization:
          Bearer <token>.'
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      expires_at:
        example: "2026-04-24T07:15:00Z"
        type: string
      refresh_token:
        description: RefreshToken is an opaque token used to obtain a new access token.
          Treat as secret.
        example: b3c4d5e6f7a8b9c0d1e2f3a4b5c6d7e8...
        type: string
      token_type:
        example: Bearer
        type: string
      user:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.UserInfo'
    type: object
...
  github_com_vndc_backend_internal_models.UserInfo:
    properties:
      avatar_uri:
        example: https://cdn.example.com/avatars/alice.png
        type: string
      bio:
        example: Blockchain enthusiast
        type: string
      country:
        example: VN
        type: string
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      email:
        example: alice@example.com
        type: string
      email_verified:
        example: true
        type: boolean
      full_name:
        example: Alice Nguyen
        type: string
      id:
        example: 65f1a2b3c4d5e6f7a8b9c0d1
        type: string
      kyc_documents:
        items:
          $ref: '#/definitions/github_com_vndc_backend_internal_models.KYCDocumentInfo'
        type: array
      kyc_level:
        example: 2
        type: integer
      kyc_status:
        example: VERIFIED
        type: string
      language:
        example: vi
        type: string
      last_login_at:
        type: string
      login_count:
        example: 42
        type: integer
      metadata:
        additionalProperties: {}
        type: object
      referral_code:
        example: ALICE2026
        type: string
      roles:
        example:
        - '["USER"]'
        items:
          type: string
        type: array
      status:
        example: ACTIVE
        type: string
      timezone:
        example: Asia/Ho_Chi_Minh
        type: string
      two_factor_enabled:
        example: false
        type: boolean
      updated_at:
        example: "2026-04-24T07:00:00Z"
        type: string
      username:
        example: alice
        type: string
      wallet_address:
        example: 0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: dev@vndc.io
    name: VNDC Platform Team
  description: |-
    Off-chain backend for the VNDC token platform.

    ## Authentication
    The API uses **wallet-based authentication** (Sign-In With Ethereum / EIP-191).
    All protected endpoints require an `Authorization: Bearer <access_token>` header.

    ### Login flow
    1. `GET /v1/auth/challenge?wallet=0x…`  — obtain a nonce-embedded message
    2. Sign the message with your wallet (personal_sign / EIP-191)
    3. `POST /v1/auth/login` — submit the signature; receive a JWT pair
    4. *(if 2FA is enabled)* `POST /v1/auth/2fa/complete` — verify TOTP code
    5. Pass the `access_token` as `Authorization: Bearer <token>` on every request
    6. `POST /v1/auth/refresh` — rotate tokens before access token expires (15 min)

    ## Amount encoding
    All token amounts are expressed in **wei** as **decimal strings** — never floats.
    Example: `"1000000000000000000"` = 1 VNDC (18 decimals).
  license:
    name: Proprietary
    url: https://vndc.io/terms
  title: VNDC Backend API
  version: 1.0.0
paths:
  /auth/2fa/complete:
    post:
      consumes:
      - application/json
      description: |-
        Verifies the TOTP code (6 digits) or a backup code (8 chars) and issues the full JWT pair.
        Requires the `temp_token` returned by POST /auth/login when `requires_2fa` is true.
      parameters:
      - description: Temp token + TOTP or backup code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_vndc_backend_internal_models.Complete2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: Full JWT pair issued
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.LoginResponse'
        "400":
          description: Missing fields
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "401":
          description: Expired session or wrong code
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
      summary: Complete two-factor authentication
      tags:
      - Auth
  /auth/2fa/disable:
    post:
      consumes:
      - application/json
      description: |-
        Disables TOTP on the account after verifying the current code or a backup code.
        This also clears the TOTP secret and all remaining backup codes.
      parameters:
      - description: 6-digit TOTP code or 8-char backup code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_vndc_backend_internal_models.Disable2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: 2FA disabled successfully
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.MessageResponse'
        "400":
          description: 2FA is not enabled
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "401":
          description: Invalid code
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deactivate two-factor authentication
      tags:
      - Auth
  /auth/2fa/enable:
    post:
      consumes:
      - application/json
      description: |-
        Confirms the TOTP secret by verifying a live authenticator code, then enables 2FA on the account.
        Must be called after POST /auth/2fa/setup.
      parameters:
      - description: Current 6-digit TOTP code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_vndc_backend_internal_models.Enable2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: 2FA enabled successfully
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.MessageResponse'
        "400":
          description: setup not called / missing code
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "401":
          description: Invalid TOTP code
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Activate two-factor authentication
      tags:
      - Auth
  /auth/2fa/setup:
    post:
      description: |-
        Creates a TOTP secret and 8 one-time backup codes for the authenticated user.
        **2FA is NOT active until the user calls POST /auth/2fa/enable** with a valid code.

        The `otp_auth_uri` field should be encoded as a QR code for the user to scan
        with an authenticator app (Google Authenticator, Authy, etc.).

        ⚠️ Backup codes are shown **once** — the user must store them securely.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.Setup2FAResponse'
        "401":
          description: Missing or invalid Bearer token
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "409":
          description: 2FA is already enabled
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Generate a new TOTP secret
      tags:
      - Auth
  /auth/challenge:
    get:
      description: |-
        Generates a random nonce and returns a pre-formatted EIP-191 message
        the wallet must sign. The nonce expires in 5 minutes.
        **Step 1** of the login flow — call before POST /auth/login.
      parameters:
      - description: Checksummed Ethereum wallet address
        example: 0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045
        in: query
        name: wallet
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ChallengeResponse'
        "400":
          description: wallet query param is missing
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
      summary: Request a SIWE sign-in challenge
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: |-
        Verifies the EIP-191 personal_sign signature against the challenge nonce.
        Returns a full JWT pair **or** a temp_token when 2FA is enabled.

        | Case | Response |
        |------|----------|
        | 2FA disabled | `LoginResponse` with `access_token` + `refresh_token` |
        | 2FA enabled  | `Login2FAResponse` with `requires_2fa: true` + `temp_token` → call `/auth/2fa/complete` |
      parameters:
      - description: Wallet address + EIP-191 signature
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_vndc_backend_internal_models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 2FA required — use temp_token
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.Login2FAResponse'
        "400":
          description: Missing or invalid fields
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "401":
          description: Signature mismatch / expired nonce / locked account
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
      summary: Sign in with Ethereum wallet (SIWE)
      tags:
      - Auth
  /auth/logout:
    post:
      description: |-
        Blacklists the current access token in Redis and revokes the active session.
        The token is rejected on subsequent requests even before its natural expiry.
      produces:
      - application/json
      responses:
        "204":
          description: Session revoked — no content
        "401":
          description: Missing or invalid Bearer token
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Log out current session
      tags:
      - Auth
  /auth/logout-all:
    post:
      description: |-
        Blacklists the current access token and revokes **all** active sessions for this user.
        Use this when a device is lost or a security breach is suspected.
      produces:
      - application/json
      responses:
        "204":
          description: All sessions revoked — no content
        "401":
          description: Missing or invalid Bearer token
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Log out all sessions
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: |-
        Exchanges a valid refresh token for a new access + refresh token pair.
        The old refresh token is **immediately invalidated** (rotation). Refresh tokens are valid for 7 days.
      parameters:
      - description: Opaque refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_vndc_backend_internal_models.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: New JWT pair
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.LoginResponse'
        "400":
          description: Missing body
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "401":
          description: Invalid / expired / already-rotated refresh token
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
      summary: Rotate the JWT token pair
      tags:
      - Auth
  /auth/sessions:
    get:
      description: |-
        Returns all unexpired, un-revoked sessions for the authenticated user.
        Each entry includes device fingerprint, IP, and geolocation so users can
        identify and revoke sessions they do not recognise.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.SessionListResponse'
        "401":
          description: Missing or invalid Bearer token
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List active sessions
      tags:
      - Auth
  /auth/sessions/{id}:
    delete:
      description: |-
        Terminates the session identified by `:id`. Users can only revoke their own sessions.
        Use this to remotely sign out a specific device.
      parameters:
      - description: Session ID (UUID)
        example: 9f8e7d6c-5b4a-3210-fedc-ba9876543210
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Session revoked — no content
        "401":
          description: Missing or invalid Bearer token
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "403":
          description: Session belongs to another user
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Revoke a specific session
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: `Bearer <token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...